internal/ui: move peer info stuff into `peerpage.go`
