+++ internal/ui/ui.go
@@ -18,7 +18,6 @@ import (
 	"github.com/diamondburned/gotk4/pkg/glib/v2"
 	"github.com/diamondburned/gotk4/pkg/gtk/v4"
 	"tailscale.com/client/tailscale/apitype"
-	"tailscale.com/ipn/ipnstate"
 	"tailscale.com/types/opt"
 )
 
@@ -73,43 +72,6 @@ func readAssetString(file string) string {
 	f := must(trayscale.Assets().Open(file))
 	must(io.Copy(&str, f))
 	return str.String()
-}
-
-func peerName(status tsutil.Status, peer *ipnstate.PeerStatus) string {
-	const maxNameLength = 30
-
-	name := tsutil.DNSOrQuoteHostname(status.Status, peer)
-	if len(name) > maxNameLength {
-		return name[:maxNameLength-3] + "..."
-	}
-	return name
-}
-
-func peerSubtitle(peer *ipnstate.PeerStatus) string {
-	if peer.ExitNode {
-		return "Current exit node"
-	}
-	if peer.ExitNodeOption {
-		return "Exit node option"
-	}
-	return ""
-}
-
-func peerIcon(peer *ipnstate.PeerStatus) string {
-	if peer.ExitNode {
-		if !peer.Online {
-			return "network-vpn-acquiring-symbolic"
-		}
-		return "network-vpn-symbolic"
-	}
-	if !peer.Online {
-		return "network-wired-offline-symbolic"
-	}
-	if peer.ExitNodeOption {
-		return "folder-remote-symbolic"
-	}
-
-	return "network-wired-symbolic"
 }
 
 func boolIcon(v bool) string {
